+++ src/cli/commands/project/index.ts
@@ -2,6 +2,7 @@ import { SgCommand, CLI } from '../../index.js';
 import { CliUx } from '@oclif/core';
 import chalk from 'chalk';
 import { Spidergram } from '../../../index.js';
+import is from '@sindresorhus/is';
 
 export default class ProjectInfo extends SgCommand {
   static description = 'Settings and stats for the current project';
@@ -26,6 +27,10 @@ export default class ProjectInfo extends SgCommand {
     if (Spidergram.status.configFile) {
       this.log(`${chalk.bold('Custom settings:')}`);
       CliUx.ux.styledJSON(sg.rawConfig);
+    }
+    if (!is.emptyObject(Spidergram.status.env)) {
+      this.log(`${chalk.bold('Environment variables:')}`);
+      CliUx.ux.styledJSON(Spidergram.status.env);  
     }
   }
 }